Install gems to vendor/bundle during CI build
When we build the site using GitHub Actions, we should install gems to
vendor/bundler. We already have a build step to cache that directory to
save time on future builds.

We also need to modify our cache key because of the previous
unsuccessful cache run due to a bug in the cache action.
https://github.com/actions/cache/issues/144#issuecomment-579323937

Fixes #39